requirements.txt
telethon
python-dotenv
pandas
python-multipart
natasha
argparse
loguru